+++ backend/core/domain/models/Professor.py
@@ -1,0 +1,14 @@
+"""Domain models for the school management system. It follows hexagonal
+architecture principles."""
+
+from dataclasses import dataclass
+from datetime import date
+
+
+@dataclass
+class Professor:
+    """Professor model class for domain layer."""
+    id: int | None
+    name: str
+    email: str # must be unique
+    password: str
